% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{peasy}
\alias{peasy}
\title{Search and stuff}
\usage{
peasy(query = NULL, visualize = TRUE)
}
\arguments{
\item{query}{Query terms}
}
\description{
Search and stuff
}
\examples{
peasy()
peasy(query = "water")
peasy(query = "water") # choose lakes, then
peasy(query = "political")

# just get URL back
peasy(query = "political", visualize = FALSE)
}

